refactor transforms to be modularized
