+++ lib/index.js
@@ -30,6 +30,12 @@ import Plain from './serializers/plain'
 import Raw from './serializers/raw'
 
 /**
+ * Transforms.
+ */
+
+import Transforms from './transforms'
+
+/**
  * Utils.
  */
 
@@ -37,6 +43,8 @@ import findDOMNode from './utils/find-dom-node'
 
 /**
  * Export.
+ *
+ * @type {Object}
  */
 
 export {
@@ -55,6 +63,7 @@ export {
   Selection,
   State,
   Text,
+  Transforms,
   findDOMNode
 }
 
@@ -74,5 +83,6 @@ export default {
   Selection,
   State,
   Text,
+  Transforms,
   findDOMNode
 }
