+++ lib/models/node.js
@@ -6,9 +6,9 @@ import Document from './document'
 import Inline from './inline'
 import Mark from './mark'
 import Selection from './selection'
-import Transforms from './transforms'
 import Text from './text'
 import direction from 'direction'
+import isInRange from '../utils/is-in-range'
 import includes from 'lodash/includes'
 import memoize from '../utils/memoize'
 import uid from '../utils/uid'
@@ -1196,38 +1196,6 @@ memoize(Node, [
 function normalizeKey(key) {
   if (typeof key == 'string') return key
   return key.key
-}
-
-/**
- * Check if an `index` of a `text` node is in a `range`.
- *
- * @param {Number} index
- * @param {Text} text
- * @param {Selection} range
- * @return {Set} characters
- */
-
-function isInRange(index, text, range) {
-  const { startKey, startOffset, endKey, endOffset } = range
-  let matcher
-
-  if (text.key == startKey && text.key == endKey) {
-    return startOffset <= index && index < endOffset
-  } else if (text.key == startKey) {
-    return startOffset <= index
-  } else if (text.key == endKey) {
-    return index < endOffset
-  } else {
-    return true
-  }
-}
-
-/**
- * Transforms.
- */
-
-for (const key in Transforms) {
-  Node[key] = Transforms[key]
 }
 
 /**
