+++ lib/models/transform.js
@@ -1,4 +1,5 @@
 
+import Transforms from '../transforms'
 import includes from 'lodash/includes'
 import xor from 'lodash/xor'
 import { List, Record } from 'immutable'
@@ -23,39 +24,13 @@ const Step = new Record({
 })
 
 /**
- * Document range transforms.
+ * Defaults.
  */
 
-const DOCUMENT_RANGE_TRANSFORMS = [
-  'deleteAtRange',
-  'deleteBackwardAtRange',
-  'deleteForwardAtRange',
-  'insertBlockAtRange',
-  'insertFragmentAtRange',
-  'insertInlineAtRange',
-  'insertTextAtRange',
-  'addMarkAtRange',
-  'setBlockAtRange',
-  'setInlineAtRange',
-  'splitBlockAtRange',
-  'splitInlineAtRange',
-  'removeMarkAtRange',
-  'toggleMarkAtRange',
-  'unwrapBlockAtRange',
-  'unwrapInlineAtRange',
-  'wrapBlockAtRange',
-  'wrapInlineAtRange',
-  'wrapTextAtRange'
-]
-
-/**
- * Document node transforms.
- */
-
-const DOCUMENT_NODE_TRANSFORMS = [
-  'removeNodeByKey',
-  'setNodeByKey',
-]
+const DEFAULT_PROPERTIES = {
+  state: null,
+  steps: new List()
+}
 
 /**
  * Selection transforms.
@@ -77,40 +52,7 @@ const SELECTION_TRANSFORMS = [
   'moveBackward',
   'moveForward',
   'moveToOffsets',
-  'moveToRangeOf'
-]
-
-/**
- * State-level document transforms.
- */
-
-const STATE_DOCUMENT_TRANSFORMS = [
-  'delete',
-  'deleteBackward',
-  'deleteForward',
-  'insertBlock',
-  'insertFragment',
-  'insertInline',
-  'insertText',
-  'addMark',
-  'setBlock',
-  'setInline',
-  'splitBlock',
-  'splitInline',
-  'removeMark',
-  'toggleMark',
-  'unwrapBlock',
-  'unwrapInline',
-  'wrapBlock',
-  'wrapInline',
-  'wrapText'
-]
-
-/**
- * State selection transforms.
- */
-
-const STATE_SELECTION_TRANSFORMS = [
+  'moveToRangeOf',
   'collapseToEndOfNextBlock',
   'collapseToEndOfNextText',
   'collapseToEndOfPreviousBlock',
@@ -121,33 +63,6 @@ const STATE_SELECTION_TRANSFORMS = [
   'collapseToStartOfPreviousText',
   'moveTo',
 ]
-
-/**
- * All state-level transforms.
- */
-
-const STATE_TRANSFORMS = []
-  .concat(STATE_DOCUMENT_TRANSFORMS)
-  .concat(STATE_SELECTION_TRANSFORMS)
-
-/**
- * All transforms.
- */
-
-const TRANSFORMS = []
-  .concat(DOCUMENT_RANGE_TRANSFORMS)
-  .concat(DOCUMENT_NODE_TRANSFORMS)
-  .concat(SELECTION_TRANSFORMS)
-  .concat(STATE_TRANSFORMS)
-
-/**
- * Defaults.
- */
-
-const DEFAULT_PROPERTIES = {
-  state: null,
-  steps: new List()
-}
 
 /**
  * Transform.
@@ -224,37 +139,14 @@ class Transform extends new Record(DEFAULT_PROPERTIES) {
 
   applyStep(state, step) {
     const { type, args } = step
+    const transform = Transforms[type]
 
-    if (includes(DOCUMENT_RANGE_TRANSFORMS, type)) {
-      let { document, selection } = state
-      let [ range, ...rest ] = args
-      range = range.normalize(document)
-      document = document[type](range, ...rest)
-      selection = selection.normalize(document)
-      state = state.merge({ document, selection })
-      return state
+    if (!transform) {
+      throw new Error(`Unknown transform type: "${type}".`)
     }
 
-    else if (includes(DOCUMENT_NODE_TRANSFORMS, type)) {
-      let { document, selection } = state
-      document = document[type](...args)
-      selection = selection.normalize(document)
-      state = state.merge({ document, selection })
-      return state
-    }
-
-    else if (includes(SELECTION_TRANSFORMS, type)) {
-      let { document, selection } = state
-      selection = selection[type](...args)
-      selection = selection.normalize(document)
-      state = state.merge({ selection })
-      return state
-    }
-
-    else if (includes(STATE_TRANSFORMS, type)) {
-      state = state[type](...args)
-      return state
-    }
+    state = transform(state, ...args)
+    return state
   }
 
   /**
@@ -271,13 +163,7 @@ class Transform extends new Record(DEFAULT_PROPERTIES) {
     const previous = undos.peek()
 
     // If the only steps applied are selection transforms, don't snapshot.
-    const onlySelections = steps.every((step) => {
-      return (
-        includes(SELECTION_TRANSFORMS, step.type) ||
-        includes(STATE_SELECTION_TRANSFORMS, step.type)
-      )
-    })
-
+    const onlySelections = steps.every(step => includes(SELECTION_TRANSFORMS, step.type))
     if (onlySelections) return false
 
     // If there isn't a previous state, snapshot.
@@ -392,7 +278,7 @@ class Transform extends new Record(DEFAULT_PROPERTIES) {
  * Add a step-creating method for each of the transforms.
  */
 
-TRANSFORMS.forEach((type) => {
+Object.keys(Transforms).forEach((type) => {
   Transform.prototype[type] = function (...args) {
     let transform = this
     let { steps } = transform
