+++ lib/utils/is-in-range.js
@@ -1,0 +1,32 @@
+
+/**
+ * Check if an `index` of a `text` node is in a `range`.
+ *
+ * @param {Number} index
+ * @param {Text} text
+ * @param {Selection} range
+ * @return {Set} characters
+ */
+
+function isInRange(index, text, range) {
+  const { startKey, startOffset, endKey, endOffset } = range
+  let matcher
+
+  if (text.key == startKey && text.key == endKey) {
+    return startOffset <= index && index < endOffset
+  } else if (text.key == startKey) {
+    return startOffset <= index
+  } else if (text.key == endKey) {
+    return index < endOffset
+  } else {
+    return true
+  }
+}
+
+/**
+ * Export.
+ *
+ * @type {Function}
+ */
+
+export default isInRange
